Implement RenderBackend protocol, and create a threejs implementation.
A commit I'm excited for. I've isolated three.js calls to the rendering system, placed as a backend implementation of a standard interface.

Also followed my example from the input-component, and have the render function behave in a more functional way (accepts a list of entities, updates them (for threejs, this is just a straight mutation of the underlying threejs objects...), and return the new list of entities (which, for threejs has no changes).